Add feed use case

+++ core/data/api/src/commonMain/kotlin/com/brbx/data/repository/RelatedRepository.kt
@@ -1,7 +1,7 @@
 package com.brbx.data.repository
 
-import com.brbx.domain.model.paging.TrackFlow
+import com.brbx.domain.model.`typealias`.TrackFlow
 
 interface RelatedRepository {
-    fun getSimilarTracks(id: Int): TrackFlow
+    fun getSimilarTracks(id: Long): TrackFlow
 }
